+++ internal/manager/manager.go
@@ -13,9 +13,9 @@ import (
 	"time"
 
 	"github.com/stashapp/stash/internal/desktop"
+	"github.com/stashapp/stash/internal/dlna"
 	"github.com/stashapp/stash/internal/log"
 	"github.com/stashapp/stash/pkg/database"
-	"github.com/stashapp/stash/pkg/dlna"
 	"github.com/stashapp/stash/pkg/ffmpeg"
 	"github.com/stashapp/stash/pkg/job"
 	"github.com/stashapp/stash/pkg/logger"
